Support HSL colors (#679)
* refactor: split rgb parsing and convertion in two functions

This will be useful for the following as we will need a function to convert from hsl to hex. To do that, we'll need a function to convert from hsl to rgb and then from rgb to hex. So, this commit helps us reuse the code for doing the last step.

* feat: add support for hsl colors

* refactor: rename test file